Extract the `Delivery.Builder` into a top-level class.
