+++ server/src/test/java/io/spine/server/delivery/DeliveryBuilderTest.java
@@ -32,10 +32,10 @@ import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-@DisplayName("Delivery Builder should")
+@DisplayName("`DeliveryBuilder` should")
 public class DeliveryBuilderTest {
 
-    private static Delivery.Builder builder() {
+    private static DeliveryBuilder builder() {
         return Delivery.newBuilder();
     }
